gdk: unmark *_async callbacks as Send
Fixes #202
