+++ gdk4/Gir.toml
@@ -182,13 +182,12 @@ generate_builder = true
     name = "set"
     manual = true # use ToValue
     [[object.function]]
-    name = "read_async"
-    manual = true # array
+    pattern = "(read|read_texture|read_text|read_value|store)_async"
+    manual = true # don't enforce Send
     [[object.function]]
     name = "set_content"
         [object.function.return]
         bool_return_is_error = "Can't set new clipboard content"
-
 
 [[object]]
 name = "Gdk.ContentDeserializer"
